Use F_SETNOSIGPIPE on OS X
It seems like OS X delivers SIGPIPE to the whole process, instead of the generating thread, as on other UNIXes. Therefore the current code sometimes works, but mostly doesn't, since the call to sigwait doesn't happen quickly enough.

Instead of masking SIGPIPE, we use an OS X specific fcntl to disable SIGPIPE for that particular fd.